backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
google-adk>=1.0.0
google-genai>=1.14.0
google-cloud-storage>=2.18.0
google-cloud-firestore>=2.19.0
python-dotenv>=1.0.0
pydantic>=2.10.0
